refactor: PerformanceCard를 공통 컴포넌트로 이동 및 접근성·반응형 개선
- components/performance/performance-card/로 이동 및 barrel export 구성
- onClick 기반 네비게이션을 Link + href prop으로 교체
- 모바일 반응형 스타일 적용 (typo, padding, line-clamp)
