+++ src/components/performance/index.ts
@@ -1,3 +1,4 @@
+export { PerformanceCard } from './performance-card';
 export { PerformanceDeleteConfirmDialog } from './performance-delete-confirm-dialog';
 export { PerformanceRegisterButton } from './performance-register-button';
 export { RegisterModal } from './register-modal';
