+++ src/components/performance/performance-card/index.ts
@@ -1,0 +1,1 @@
+export { PerformanceCard } from './performance-card';
